% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DarkSky-package.R
\docType{package}
\name{DarkSky-package}
\alias{DarkSky}
\alias{DarkSky-package}
\title{DarkSky: R Client for the Dark Sky API}
\description{
The DarkSky package provides an R interface to the Dark Sky API
service, facilitating API calls and processing API responses into
convenient, R-friendly data structures. See
<https://darksky.net/dev/docs> for the Dark Sky API Documentation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/fascinatingfingers/darksky}
  \item Report bugs at \url{https://gitlab.com/fascinatingfingers/darksky/issues}
}

}
\keyword{internal}
