% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darksky-package.r
\docType{package}
\name{darksky}
\alias{darksky}
\alias{darksky-package}
\title{Tools to Work with the Dark Sky API}
\description{
Programmatic access to the Dark Sky API <https://darksky.net/dev/docs>,
which provides current or historical global weather conditions.
}
\author{
Bob Rudis (bob@rud.is)
}

