% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-key.r
\name{darksky_api_key}
\alias{darksky_api_key}
\title{Get or set DARKSKY_API_KEY value}
\usage{
darksky_api_key(force = FALSE)
}
\arguments{
\item{force}{force setting a new Dark Sky API key for the current environment?}
}
\value{
atomic character vector containing the Dark Sky API key
}
\description{
The API wrapper functions in this package all rely on a Dark Sky API
key residing in the environment variable \code{DARKSKY_API_KEY}. The
easiest way to accomplish this is to set it in the `\code{.Renviron}` file in your
home directory.
}
