% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_forecast.R
\name{get_forecast}
\alias{get_forecast}
\title{Get Weather Forecast From Dark Sky API}
\usage{
get_forecast()
}
\value{
Returns the Dark Sky response, parsed into an R-friendly list.
}
\description{
Get Weather Forecast From Dark Sky API
}
\seealso{
\url{https://darksky.net/dev/docs}
}
