% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-plot.r
\name{plot.darksky}
\alias{plot.darksky}
\title{Plot method for darksky objects}
\usage{
\method{plot}{darksky}(x, ..., readings = c("hourly", "minutely", "daily"))
}
\arguments{
\item{x}{a darksky x}

\item{readings}{specify which readings to plot. will plot all available by default}

\item{\dots}{ignored}
}
\value{
frame grob
}
\description{
Uses ggplot2 & grid.arrange to produce graphs for darksky objects
}
\note{
only forecast/readings components of \code{x} that have more than one observation
  will be plotted
}

