% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-print.r
\name{print.darksky}
\alias{print.darksky}
\title{Slightly more human-readable output for darksky objects}
\usage{
\method{print}{darksky}(x, ...)
}
\arguments{
\item{x}{a darksky object}

\item{...}{ignored}
}
\description{
Slightly more human-readable output for darksky objects
}
