% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{set_dark_sky_credentials}
\alias{set_dark_sky_credentials}
\alias{reset_dark_sky_credentials}
\title{Dark Sky API Authentication}
\usage{
set_dark_sky_credentials(key, lat, lon)

reset_dark_sky_credentials()
}
\arguments{
\item{key}{Dark Sky secret key}

\item{lat}{latitude}

\item{lon}{longitude}
}
\value{
Returns \code{TRUE} (invisibly) if a connection to the Hue Bridge was
  successful.
}
\description{
These functions are used primarily for their side effects -- specifically to
set (or reset) Dark Sky API authentication secrets stored in
\code{options()$DarkSky}.
}
\seealso{
\url{https://darksky.net/dev/docs}
}
