% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_helpers.R
\name{set_dark_sky_credentials}
\alias{set_dark_sky_credentials}
\alias{reset_dark_sky_credentials}
\title{Dark Sky API Authentication}
\usage{
set_dark_sky_credentials(key, lat, lon)

reset_dark_sky_credentials()
}
\arguments{
\item{key}{Dark Sky secret key}

\item{lat}{latitude (in decimal degrees). Positive north; negative south.}

\item{lon}{longitude (in decimal degrees). Positive east; negative west.}
}
\value{
Returns \code{TRUE} (invisibly) if options were successfully set or
  reset.
}
\description{
These functions are used specifically for their side effects -- namely to
set (or reset) Dark Sky API authentication secrets stored in
\code{options()$DarkSky}.
}
\seealso{
\url{https://darksky.net/dev/docs}
}
